requirements.txt
arabic-reshaper==2.1.0
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.2
django-cors-headers==3.5.0
django-import-export==2.4.0
django-simple-history==2.11.0
django-smart-selects==1.5.9
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
docopt==0.6.2
et-xmlfile==1.0.1
future==0.18.2
html5lib==1.1
idna==2.10
jdcal==1.4.1
MarkupPy==1.14
num2words==0.5.10
odfpy==1.4.1
openpyxl==3.0.5
Pillow==7.2.0
PyJWT==1.7.1
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2020.1
PyYAML==5.3.1
reportlab==3.5.53
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
tablib==2.0.0
urllib3==1.25.10
webencodings==0.5.1
xhtml2pdf==0.2.5
xlrd==1.2.0
xlwt==1.3.0
